Feedback + Player system + v++
Adding a debug Feedback system, as well as a system to get player objects with plugin specific information.
PlayerCache FPPlayer - TODO: add more stuff... its not really required rn, if nothing added, remove.
FeedbackModule - TODO: outsource the Feedback class + make it an interface and implement a TestFeedback or so. To later on switch between test and full framework